relocate sidebar/nav to own folder
