+++ src/components/BaseHeader.tsx
@@ -4,7 +4,7 @@ import { getServerAuthSession } from '@src/server/auth';
 import { ClubSearchBar } from './searchBar/ClubSearchBar';
 import { EventSearchBar } from './searchBar/EventSearchBar';
 import SignInButton from './signInButton';
-import MobileNav from './MobileNav';
+import MobileNav from './nav/MobileNav';
 import { api } from '@src/trpc/server';
 
 export const BaseHeader = async ({ children }: { children: ReactNode }) => {
