+++ src/components/nav/SidebarItems.tsx
@@ -2,7 +2,7 @@
 import { type FC, useState } from 'react';
 import { useRouter, usePathname } from 'next/navigation';
 import { IconMap, type allCats, routeMap } from '@src/constants/categories';
-import { RightChevron } from '../icons/Icons';
+import { RightChevron } from '../../icons/Icons';
 
 const SidebarItems: FC<{ cat: allCats[number] }> = ({ cat }) => {
   const Icon = IconMap[cat];
